Solution to Input and Print a Float

+++ Input and Print a Float@@6788d5012db25efec946f93a/code.c
@@ -1,6 +1,5 @@
 #include <stdio.h>
-
-float main{
+int main(){
     float number;
     scanf("%f",&number);
     printf("Ypu entered: %.2f\n ",number);
